Fix button

+++ HueControlServer/SNZB_01Message.cs
@@ -3,7 +3,7 @@
     public class SNZB_01Message
     {
         public string? action { get; set; }
-        public int battery { get; set; }
+        public float battery { get; set; }
         public int linkquality { get; set; }
         public int voltage { get; set; }
     }
